evals/transcription/pyproject.toml
[tool.poetry]
name = "transcription-evals"
version = "0.1.0"
description = "Evaluation suite for transcription services"
authors = ["i.AI <i-dot-ai-enquiries@cabinetoffice.gov.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
datasets = ">=2.0.0,<3.0.0"
evaluate = ">=0.4.0,<0.5.0"
jiwer = ">=3.0.0,<4.0.0"
soundfile = ">=0.12.0,<0.13.0"
librosa = ">=0.10.0,<0.11.0"
python-dotenv = ">=1.0.0,<2.0.0"
tqdm = ">=4.0.0,<5.0.0"
pyannote-metrics = ">=3.2.0,<4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
